iio/adc/ad9361: Fix digital interface timing issues in FIR enable mode
The FIR filter once enabled causes the interface timing to change.
It's typically not a problem if the timing margin is big enough.
However at 61.44 MSPS it causes problems on some systems.
A common clk/data delay timing configuration that serves all use cases
may not exist on these systems.
So we always run the digital tune in case the filter is enabled.
If it is disabled we restore the values from the initial calibration.

This new feature can be disabled from the device tree using:
	adi,digital-interface-tune-fir-disable;

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>