dma: xilinx: use 'reg' property to specify chan ID
On some platforms the devicetree may not be enumerated in the order
expected by drivers, causing the DMA channel IDs to be in reverse order.
For other nodes that are attaching to a channel using the channel ID,
this is problemaitic.

Enable the use of the 'reg' property in each of the channel nodes to
deterministically specify the desired channel ID, and allow other nodes
to easily point to it.

Signed-off-by: Matt Fornero <matt.fornero@mathworks.com>