dma: xilinx: Reset when halt fails
For some configurations, the halt operation may fail, resulting in an
unuable DMA channel. If specified in the device tree, fall back to
a channel reset in these cases to enable continued use of the DMA
channel.

Note that the reset is device-wide, so this should be taken into
consideration when enabling this option.

Signed-off-by: Matthew Fornero <mfornero@mathworks.com>
Signed-off-by: Matt Fornero <matt.fornero@mathworks.com>