dma: xilinx: Allow halt operation to reset DMA
For some configurations, the halt operation may fail, resulting in an
unusable DMA channel. For other configurations, the halt operation may
succeed, but the DMA engine may be left in a corrupt state due to
activity between the DMA engine and a streaming device as the channel
shuts down.

Add device tree options to enable the halt operation to reset the DMA
engine either unconditionally or if the halt attempt fails.

Signed-off-by: Matt Fornero <matt.fornero@mathworks.com>