[media] adv7511: Prevent CSC config changes
For platforms that use a YCbCr input and require a RGB output, the CSC
must be enabed and configured with the appropriate coefficients,
generally through the device tree.

However, changes to the timing or other modes can trigger a call to
adv7511_set_rgb_quantization_mode, which will either disable the CSC or
set the coefficients to those suitable for RGB range conversion,
resulting in bad video output.

Add a device tree flag, csc-locked, to prevent calls to
adv7511_set_rgb_quantization_mode from changing the CSC configuration.

Signed-off-by: Matt Fornero <matt.fornero@mathworks.com>