dma: xilinx: Add DT entry for len register width
The width of the length register is a configurable option in the DMA IP
core, but is currently hard coded in the driver to 23 bits. If an
instance of the IP uses a narrower length width, the driver will fail to
work for transactions larger than a single physical segment.

Add a device tree entry "xlnx,lenwidth" to specify the width of this
register in hardware. The value will be used to determine how
transactions are broken up into physical segments. If not specified, the
existing default values will be used.

Signed-off-by: Matt Fornero <matt.fornero@mathworks.com>