dma: xilinx: Support partial transfers
If a partial transfer has occured (through early assertion of TLAST),
all of the segments in a descriptor chain may not be marked complete,
so simply testing the last segment's status bit is not sufficient.

Instead, *if* the last segment is not done *and* it's a S2MM transfer,
step through each segment in the chain to check for a completion.

Signed-off-by: Matt Fornero <matt.fornero@mathworks.com>