iio: adrv9002: Adapt driver to the new API version
This patch changes the default profiles in order to be valid on the new API
version. Also to mention that the CMOS stream binary was removed since it
was identical to the one generated for the LVDS profile.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>
(cherry picked from commit c6ee95eb75dcc002ebffeb08bbd32cf220e4a6d4)