iio: adrv9002: Export TDD clocks
The TDD clock is given by the product between the dds rate and the
interface clock. As the adrv9002 driver is the one having all of this
info, we do all the calculations in the driver and export a new tdd
clock to be used by the tdd core driver.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>
(cherry picked from commit bad9e7349dc3795a8bfe6386e35b800aa904acae)